tests/integration: Change default AVAL machines to speed up tests
Since AVAL lock a device, finish the tests and then release the device, it's
not possible to run AVAL tests with same device in parallel. This commit
change the default machines to run only one job per machine, this way it's
possible to speed up the tests and cover more machines.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>