tests/integration: Fix build and bundle commands failing in WSL
Commit 0578e83bef55c92aa0f1b452873c2db3110f423e added the "-p 8080:8080
configuration to the TorizonCore Builder Docker commands on WSL as a way to
make the ostree serve command work. However, this configuration causes the
build, bundle, and bundle-registry commands to fail.

While "-p 8080:8080" does allow ostree serve to work, it is also possible to
use the "--network=host configuration", similar to Linux, provided the Windows
Firewall is configured correctly. Using "--network=host" fixes the failures in
the build and bundle commands.

Related-to: TCB-507

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>